Fix database errors when updating EZID ARK records
Update arkimedes.upload to use different functions for adding a new
record to the database when minting an ARK and for updating an existing
database record when updating an existing ARK. In arkimedes.db add
update_db_record function and add optional session argument to find
functions, so that functions that need to update the database can make
queries and make changes to the results. In arkimedes.ezid add
anvl_to_dict function to convert a single-record ANVL string to a
dictionary.

Unrelated to above changes: in arkimedes.db add to_anvl() method to Ark
class, which returns an ANVL-formatted string of an instances values;
add stub function sync_db.