Fix bugs; add and refactor database functionality
Fix bug in Ark.from_anvl() so it handles booleans represented as strings
properly. Add dump_db() function to create database backups in ANVL
format. Rename load_batch_download_file_into_db() as
load_anvl_file_into_db() and change functionality so it can handle any
ANVL file that matches or is a subset of the arkimedes.db.Ark schema.
The function now creates a database if one was not already created and
has a verbose mode (on by default) to print the number of the ARK record
being added to the database.

Add function load_anvl_as_str in arkimedes.ezid.